% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{browser}
\alias{browser}
\alias{provBookR.prov.browse}
\alias{prov.browser}
\alias{UI.user.input.quit}
\alias{UI.user.input.help}
\alias{UI.user.input.AO}
\alias{UI.user.input.BO}
\alias{UI.user.input.S}
\alias{UI.user.input.DB}
\alias{UI.user.input.C}
\alias{UI.user.input.BG}
\title{Provenance browser}
\usage{
provBookR.prov.browse(file.name)

prov.browser(file.name)

UI.user.input.quit()

UI.user.input.help()

UI.user.input.AO(prov.json, data.object.name, history)

UI.user.input.BO(prov.json, data.object.name, history)

UI.user.input.S(file.name)

UI.user.input.DB(file.name)

UI.user.input.C(file.name, data.object.name)

UI.user.input.BG(prov.json, data.object.name)
}
\arguments{
\item{file.name}{name of provenance file}

\item{prov.json}{provenance file}

\item{data.object.name}{name of data object}

\item{history}{the provenance sequence}
}
\value{
provBookR browser

provbook options

forward lineage

backward lineage

summarized provenance

provDebugR browser interface

cleaned .R script

provbook
}
\description{
browser
}
