% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentation.R
\name{presentation enables the viewing of provenance as a data a frame}
\alias{presentation enables the viewing of provenance as a data a frame}
\alias{provBookR.prov.presentation}
\title{provenance presenter}
\usage{
provBookR.prov.presentation(prov.df)
}
\arguments{
\item{prov.df}{data frame containing provenance}
}
\description{
Provenance presentation for provBookR
}
