% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{processing}
\alias{processing}
\alias{provBookR.prov.processing}
\title{Provenance browser for provBookR}
\usage{
provBookR.prov.processing(parsed.prov.df)
}
\arguments{
\item{parsed.prov.df}{the parsed data frame provenance}
}
\description{
Provenance browser for provBookR
}
