% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recording.R
\name{prov.recording}
\alias{prov.recording}
\title{provenance recorder}
\usage{
prov.recording(file.name)
}
\arguments{
\item{file.name}{name of provenance file}
}
\description{
Provenance recorder for provBookR
}
\details{
\code{prov.recording}
}
