% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provBook.R
\name{provBook}
\alias{provBook}
\alias{provBookR}
\title{Main provBookR interface}
\usage{
provBookR(file.name, mode)
}
\arguments{
\item{file.name}{provenance file name}

\item{mode}{type of browsing}
}
\value{
full provBookR browser or enable lite operation
}
\description{
Main provBookR interface
}
