% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querying.R
\name{querying the provenance for specific values}
\alias{querying the provenance for specific values}
\alias{provBookR.prov.querying}
\title{Provenance querying for provBookR}
\usage{
provBookR.prov.querying(prov.json, data.object.name, history)
}
\arguments{
\item{prov.json}{provenance file}

\item{data.object.name}{name of the data object file}

\item{history}{object provenance}
}
\description{
Provenance querying for provBookR
}
